Install commands are united in wrong way.
The correct way of executing one command after another is using double ampersand (&&), which stands for AND, so if the first with SUCCESS, the second will be executed. If not, the second command will be ignored.
The way it was - single ampersand (&) - the process was going to background, because that's what a single ampersand does.